Move template AST logic for classes/attributes to transformTemplate()
