+++ lib/transform.js
@@ -52,19 +52,32 @@ function transform(source, template, options = {}) {
   let type = checkComponentType(root);
   let result;
 
-  switch (type) {
-    case 'classic':
-      result = transformClassicComponent(root, options);
-      break;
-    default:
-      throw new Error(
-        `Unsupported component type. Only classic components (\`Component.extend({ ... }\`) are supported currently.`
-      );
+  if (type === 'classic') {
+    result = transformClassicComponent(root, options);
+  } else {
+    throw new SilentError(
+      `Unsupported component type. Only classic components (\`Component.extend({ ... }\`) are supported currently.`
+    );
   }
 
   if (result) {
-    let { newSource, attrs, tagName } = result;
-    let newTemplate = transformTemplate(template, tagName, attrs);
+    let {
+      newSource,
+      tagName,
+      elementId,
+      classNames,
+      classNameBindings,
+      attributeBindings,
+    } = result;
+    let newTemplate = transformTemplate(
+      template,
+      tagName,
+      elementId,
+      classNames,
+      classNameBindings,
+      attributeBindings,
+      options
+    );
 
     return { source: newSource, template: newTemplate, tagName };
   }
